pyproject.toml
[tool.poetry]
name = "test-especialist"
version = "0.1.0"
description = ""
authors = ["holynacll <alexandrecurylima@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.1"}
pandas = "^2.2.2"
openpyxl = "^3.1.5"
pymupdf = "^1.24.9"
google-generativeai = "^0.7.2"
black = "^24.8.0"
spacy = "^3.7.6"
unidecode = "^1.3.8"
ipython = "^8.26.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
